+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * added `modal` and `modal_trigger` for Bootstrap modal dialogs
 * added `modal_alert` and `modal_confirm` which work similarly to their Javascript counterparts
 * added `flash_messages` helper for flash support
+* added `bs_table_tag` for consistent Bootstrap tables
 
 ## v 0.0.11